pyproject.toml
[project]
name = "ml-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.14.0",
    "openai>=1.60.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "types-requests>=2.32.0.20241016",
    "watchfiles>=1.0.4",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.0",
    "mypy>=1.14.1",
    "pytest-sugar>=1.0.0",
    "ruff>=0.9.4",
]
